Introduce lib/backoff.js and move sendSafelyRetryable to it
backoff.js includes the backoffRetrier function, which creates a
higher-order function that will retry any async function passed to it
until it reaches a max attempt count or until it succeeds.
backoffRetrier allows customizing which errors result in a retry vs an
immediate termination.
